requirements.txt
pandas==2.1.0
numpy==1.24.3
psycopg2-binary==2.9.7
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.5.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
scipy==1.11.3
scikit-learn==1.3.0
seaborn==0.12.2
matplotlib==3.7.2
pytest==7.4.2
black==23.7.0
flake8==6.1.0
faker==19.6.2
dash-bootstrap-components==1.5.0
plotly-express==0.4.1
dash-table==5.0.0